compontizing companies and new company
